refactor(core): unify naming convention for files
